Firstly, I need to work on a scene system - raylib-style won't work as nice as intended
