+++ graphics/CMakeLists.txt
@@ -7,6 +7,7 @@ add_library(graphics STATIC
 	models.c
 	renderer.c
 	shaders.c
+	scene.c
 )
 
 target_include_directories(graphics PUBLIC .)
